say restricted to owner

+++ commands/say.js
@@ -1,8 +1,8 @@
 module.exports = {
   name: 'say',
   description: 'Say!',
-  execute(message, args) {
-    // if (message.author.id !== client.config.ownerID) return;
+  execute(client, message, args) {
+    if (message.author.id !== client.config.ownerID) return;
     message.channel.send(args.join(' '));
     message.delete();
   },
